+++ common/src/main/java/com/chefmooon/frightsdelight/common/registry/FrightsDelightBlockEntities.java
@@ -6,6 +6,7 @@ import net.minecraft.resources.ResourceLocation;
 public class FrightsDelightBlockEntities {
     public static final ResourceLocation LOLLIPOP_MOLD = blockEntity("lollipop_mold");
     public static final ResourceLocation RING_CANDY_MOLD = blockEntity("ring_candy_mold");
+    public static final ResourceLocation CANDY_BASKET = blockEntity("candy_basket");
     private static ResourceLocation blockEntity(String string) {
         return TextUtils.res(string);
     }