+++ fabric/src/main/java/com/chefmooon/frightsdelight/client/event/fabric/ClientSetupEventsImpl.java
@@ -33,6 +33,15 @@ public class ClientSetupEventsImpl {
                 FrightsDelightBlocksImpl.LOLLIPOP_MOLD,
                 FrightsDelightBlocksImpl.RING_CANDY_MOLD
         );
+
+        BlockRenderLayerMap.INSTANCE.putBlocks(RenderType.cutout(),
+                FrightsDelightBlocksImpl.PUMPKIN_CANDY_BASKET,
+                FrightsDelightBlocksImpl.MELON_CANDY_BASKET
+//                FrightsDelightBlocksImpl.CREEPER_CANDY_BASKET,
+//                FrightsDelightBlocksImpl.SKELETON_CANDY_BASKET,
+//                FrightsDelightBlocksImpl.WITHER_SKELETON_CANDY_BASKET,
+//                FrightsDelightBlocksImpl.ZOMBIE_CANDY_BASKET
+        );
     }
 
     public static void registerParticles() {