+++ fabric/src/main/java/com/chefmooon/frightsdelight/common/registry/fabric/FrightsDelightBlockEntitiesImpl.java
@@ -1,5 +1,6 @@
 package com.chefmooon.frightsdelight.common.registry.fabric;
 
+import com.chefmooon.frightsdelight.common.block.entity.CandyBasketBlockEntity;
 import com.chefmooon.frightsdelight.common.block.entity.LollipopMoldBlockEntity;
 import com.chefmooon.frightsdelight.common.block.entity.RingCandyMoldBlockEntity;
 import com.chefmooon.frightsdelight.common.registry.FrightsDelightBlockEntities;
@@ -21,6 +22,15 @@ public class FrightsDelightBlockEntitiesImpl {
                     FrightsDelightBlocksImpl.RING_CANDY_MOLD
             ));
 
+    public static final BlockEntityType<CandyBasketBlockEntity> CANDY_BASKET = registerBlockEntity(FrightsDelightBlockEntities.CANDY_BASKET,
+            BlockEntityType.Builder.of(CandyBasketBlockEntity::new,
+                    FrightsDelightBlocksImpl.PUMPKIN_CANDY_BASKET,
+                    FrightsDelightBlocksImpl.MELON_CANDY_BASKET
+//                    FrightsDelightBlocksImpl.CREEPER_CANDY_BASKET,
+//                    FrightsDelightBlocksImpl.SKELETON_CANDY_BASKET,
+//                    FrightsDelightBlocksImpl.WITHER_SKELETON_CANDY_BASKET,
+//                    FrightsDelightBlocksImpl.ZOMBIE_CANDY_BASKET
+            ));
 
 
     public static <T extends BlockEntity> BlockEntityType<T> registerBlockEntity(ResourceLocation location, BlockEntityType.Builder<T> builder) {