+++ fabric/src/main/java/com/chefmooon/frightsdelight/common/utility/fabric/FrightsDelightTextureSlots.java
@@ -8,4 +8,6 @@ public class FrightsDelightTextureSlots {
     public static final TextureSlot INSIDE = TextureSlot.create("inside");
     public static final TextureSlot CUP = TextureSlot.create("cup");
     public static final TextureSlot SYRUP = TextureSlot.create("syrup");
+    public static final TextureSlot BASKET = TextureSlot.create("basket");
+    public static final TextureSlot HANDLE = TextureSlot.create("handle");
 }