+++ fabric/src/main/java/com/chefmooon/frightsdelight/data/fabric/LootTableGenerator.java
@@ -11,6 +11,7 @@ import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTableProvider;
 import net.minecraft.advancements.critereon.ItemPredicate;
 import net.minecraft.advancements.critereon.StatePropertiesPredicate;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.core.component.DataComponents;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.models.model.ModelLocationUtils;
 import net.minecraft.world.item.Item;
@@ -21,6 +22,7 @@ import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.LootTable;
 import net.minecraft.world.level.storage.loot.entries.LootItem;
 import net.minecraft.world.level.storage.loot.functions.ApplyBonusCount;
+import net.minecraft.world.level.storage.loot.functions.CopyComponentsFunction;
 import net.minecraft.world.level.storage.loot.functions.SetItemCountFunction;
 import net.minecraft.world.level.storage.loot.predicates.LootItemBlockStatePropertyCondition;
 import net.minecraft.world.level.storage.loot.predicates.MatchTool;
@@ -38,6 +40,14 @@ public class LootTableGenerator extends FabricBlockLootTableProvider {
 
     @Override
     public void generate() {
+
+        this.createCandyBasketDrops(FrightsDelightBlocksImpl.PUMPKIN_CANDY_BASKET);
+        this.createCandyBasketDrops(FrightsDelightBlocksImpl.MELON_CANDY_BASKET);
+
+//        this.createCandyBasketDrops(FrightsDelightBlocksImpl.CREEPER_CANDY_BASKET);
+//        this.createCandyBasketDrops(FrightsDelightBlocksImpl.SKELETON_CANDY_BASKET);
+//        this.createCandyBasketDrops(FrightsDelightBlocksImpl.WITHER_SKELETON_CANDY_BASKET);
+//        this.createCandyBasketDrops(FrightsDelightBlocksImpl.ZOMBIE_CANDY_BASKET);
 
         this.dropSelf(FrightsDelightBlocksImpl.FLESH_CRATE);
         this.dropSelf(FrightsDelightBlocksImpl.BONE_CRATE);
@@ -152,5 +162,13 @@ public class LootTableGenerator extends FabricBlockLootTableProvider {
                                 .setProperties(StatePropertiesPredicate.Builder.properties().hasProperty(PieBlock.BITES, 3))
                                 .and(MatchTool.toolMatches(ItemPredicate.Builder.item().of(CommonTags.C_TOOLS_KNIFE)))))
         ));
+    }
+
+    private void createCandyBasketDrops(Block block) {
+        this.add(block, this.applyExplosionDecay(block, LootTable.lootTable()
+                .withPool(LootPool.lootPool().setRolls(ConstantValue.exactly(1.0F))
+                        .add(LootItem.lootTableItem(block)
+                                .apply(CopyComponentsFunction.copyComponents(CopyComponentsFunction.Source.BLOCK_ENTITY).include(DataComponents.CONTAINER).include(DataComponents.CUSTOM_NAME)))))
+        );
     }
 }