+++ fabric/src/main/java/com/chefmooon/frightsdelight/data/fabric/TranslationGenerator.java
@@ -149,7 +149,17 @@ public class TranslationGenerator extends FabricLanguageProvider {
         translationBuilder.add(FrightsDelightItemsImpl.WITHER_BERRY_RING_CANDY, "Wither Berry Ring Candy");
         translationBuilder.add(FrightsDelightItemsImpl.COBWEB_RING_CANDY, "Cobweb Ring Candy");
 
+        translationBuilder.add(FrightsDelightItemsImpl.PUMPKIN_CANDY_BASKET, "Pumpkin Candy Basket");
+        translationBuilder.add(FrightsDelightItemsImpl.MELON_CANDY_BASKET, "Melon Candy Basket");
+
+//        translationBuilder.add(FrightsDelightItemsImpl.CREEPER_CANDY_BASKET, "Creeper Candy Basket");
+//        translationBuilder.add(FrightsDelightItemsImpl.SKELETON_CANDY_BASKET, "Skeleton Candy Basket");
+//        translationBuilder.add(FrightsDelightItemsImpl.WITHER_SKELETON_CANDY_BASKET, "Wither Skeleton Candy Basket");
+//        translationBuilder.add(FrightsDelightItemsImpl.ZOMBIE_CANDY_BASKET, "Zombie Candy Basket");
+
         translationBuilder.add(FrightsDelightBlocksImpl.LOLLIPOP_MOLD.getDescriptionId() + ".bone_shard_requirement", "Requires 4 Bone Shards");
+
+        translationBuilder.add("container.candy_basket", "Candy Basket");
 
         translationBuilder.add(EFFECT + ".chills", "Chills");
         translationBuilder.add(EFFECT + ".cobwebbed", "Cobwebbed");
@@ -198,6 +208,11 @@ public class TranslationGenerator extends FabricLanguageProvider {
         translationBuilder.add(SUBTITLE + ".block.candy_mold.remove", "Bone Shard(s) Removed");
         translationBuilder.add(SUBTITLE + ".block.candy_mold.collect", "Candy Taken");
         translationBuilder.add(SUBTITLE + ".block.candy_mold.harden", "Candy Hardened");
+
+        translationBuilder.add(SUBTITLE + ".block.candy_basket.add", "Candy Added");
+        translationBuilder.add(SUBTITLE + ".block.candy_basket.remove", "Candy Removed");
+        translationBuilder.add(SUBTITLE + ".block.candy_basket.shuffle", "Candy Shuffled"); // change to Mixed?
+        translationBuilder.add(SUBTITLE + ".block.candy_basket.empty", "Candy Basket Empty");
 
         translationBuilder.add(DEATH + ".soul_berry_bush", "%1$s was pricked to death by a Soul Berry Bush");
         translationBuilder.add(DEATH + ".soul_berry_bush.player", "%1$s was sent to a prickly grave by %2$s");