+++ neoforge/src/main/java/com/chefmooon/frightsdelight/common/registry/neoforge/FrightsDelightBlockEntityImpl.java
@@ -1,6 +1,7 @@
 package com.chefmooon.frightsdelight.common.registry.neoforge;
 
 import com.chefmooon.frightsdelight.FrightsDelight;
+import com.chefmooon.frightsdelight.common.block.entity.CandyBasketBlockEntity;
 import com.chefmooon.frightsdelight.common.block.entity.LollipopMoldBlockEntity;
 import com.chefmooon.frightsdelight.common.block.entity.RingCandyMoldBlockEntity;
 import com.chefmooon.frightsdelight.common.registry.FrightsDelightBlockEntities;
@@ -22,6 +23,16 @@ public class FrightsDelightBlockEntityImpl {
     public static final Supplier<BlockEntityType<RingCandyMoldBlockEntity>> RING_CANDY_MOLD = BLOCK_ENTITIES.register(FrightsDelightBlockEntities.RING_CANDY_MOLD.getPath(),
             () -> BlockEntityType.Builder.of(RingCandyMoldBlockEntity::new,
                     FrightsDelightBlocksImpl.RING_CANDY_MOLD.get()
+            ).build(null));
+
+    public static final Supplier<BlockEntityType<CandyBasketBlockEntity>> CANDY_BASKET = BLOCK_ENTITIES.register(FrightsDelightBlockEntities.CANDY_BASKET.getPath(),
+            () -> BlockEntityType.Builder.of(CandyBasketBlockEntity::new,
+                    FrightsDelightBlocksImpl.PUMPKIN_CANDY_BASKET.get(),
+                    FrightsDelightBlocksImpl.MELON_CANDY_BASKET.get()
+//                    FrightsDelightBlocksImpl.CREEPER_CANDY_BASKET.get(),
+//                    FrightsDelightBlocksImpl.SKELETON_CANDY_BASKET.get(),
+//                    FrightsDelightBlocksImpl.WITHER_SKELETON_CANDY_BASKET.get(),
+//                    FrightsDelightBlocksImpl.ZOMBIE_CANDY_BASKET.get()
             ).build(null));
 
     public static void register(IEventBus eventBus) {